フォルダ構成を変更、Rakefileでテスト自動化
