+++ 06.bowling_object/Rakefile
@@ -1,0 +1,9 @@
+require 'rake/testtask'
+
+
+
+Rake::TestTask.new do |t|
+  t.pattern = 'test/**/*_test.rb'
+end
+
+task default: :test
