+++ 06.bowling_object/libs/frame.rb
