+++ 06.bowling_object/libs/game.rb
