+++ 06.bowling_object/libs/main.rb
