+++ 06.bowling_object/libs/shot.rb
